+++ casbin_pymongo_adapter/__init__.py
@@ -1,1 +1,9 @@
 from .adapter import Adapter
+from ._filter import Filter
+from ._rule import CasbinRule
+
+__all__ = [
+    "Adapter",
+    "Filter",
+    "CasbinRule",
+]
